refactor(monitor): Extract timeFrame utils to break HMR boundary
ServicesView exported non-component values (getLoopbackInterval,
yAxisTickFormat) alongside a React component, which disables React Fast
Refresh for the whole file. Moving them to timeFrameUtils.ts restores
HMR for ServicesView during development.
